Temp fix zng-ext-font Android build
